[refactor] component 관리 시작 - githubLoginSutton - 기존 githubLogin 을 Button 컴포넌트로 분류하고 컴포넌트 안에 귀속시킴
